Add proxy routes
